gmscompat: Make CTS/Play Integrity pass again
The logic behind CTS and Play Integrity has been updated today it now
checks the product and model names against the fingerprint and if
they do not match the CTS profile will fail.

Also switch to a newer FP from Pixel 2 while we are at it.

Squashed also with the following commit:

Author: Dyneteve <dyneteve@hentaios.com>
Date:   Tue Aug 23 18:57:05 2022 +0200

    gmscompat: Apply the SafetyNet workaround to Play Store

    Play Store is used for the new Play Integrity API, extend the hack
    to it as well

    Test: Device Integrity and Basic Integrity passes.

    Signed-off-by: Dyneteve <dyneteve@hentaios.com>
    Change-Id: Id607cdff0b902f285a6c1b769c0a4ee4202842b1

Test: Boot, check for CTS and Play Integrity
Change-Id: I089d5ef935bba40338e10c795ea7d181103ffd15
Signed-off-by: Dyneteve <dyneteve@hentaios.com>